рефакторинг файла утилит
